Add edit/delete, Targets page, half-step scores, and report export
- Assessments list: Edit (unpublish) and Delete actions; switched to client
  fetch with include_counts=1 query so counts stay live.
- Score page: header now shows Edit (unpublish) when published plus a Report
  link; numeric score type lets selector handle half steps.
- New /targets page mirrors the scoring UI for setting target maturity per
  subcategory, with per-function quick-set buttons.
- MaturitySelector: full 1-5 buttons keep their rich tooltips; a smaller row
  of 1.5/2.5/3.5/4.5 buttons offers half-step selection without tooltips.
- Schema: score and target_score widened to REAL to formalise half-step
  storage (SQLite already permitted it).
- Report view at /assessments/[id]/report (printable, shareable URL, Copy
  Link button) plus CSV export endpoint at
  /api/assessments/[id]/export?format=csv (also supports format=json).

+++ src/lib/db.ts
@@ -53,7 +53,7 @@ export function getDb(): Database.Database {
       id INTEGER PRIMARY KEY AUTOINCREMENT,
       assessment_id TEXT NOT NULL REFERENCES assessments(id) ON DELETE CASCADE,
       control_id TEXT NOT NULL REFERENCES controls(id),
-      score INTEGER CHECK(score BETWEEN 1 AND 5),
+      score REAL CHECK(score BETWEEN 1 AND 5),
       rationale TEXT,
       created_at TEXT DEFAULT (datetime('now')),
       updated_at TEXT DEFAULT (datetime('now')),
@@ -64,7 +64,7 @@ export function getDb(): Database.Database {
       id INTEGER PRIMARY KEY AUTOINCREMENT,
       framework_id INTEGER NOT NULL REFERENCES frameworks(id),
       control_id TEXT NOT NULL REFERENCES controls(id),
-      target_score INTEGER NOT NULL CHECK(target_score BETWEEN 1 AND 5),
+      target_score REAL NOT NULL CHECK(target_score BETWEEN 1 AND 5),
       notes TEXT,
       updated_at TEXT DEFAULT (datetime('now')),
       UNIQUE(framework_id, control_id)
